--- pkg/util/util.go
@@ -1,42 +1,0 @@
-// Copyright 2023 OnMetal authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package util
-
-import (
-	"errors"
-	"fmt"
-	"strings"
-)
-
-const (
-	CloudProviderName = "onmetal"
-)
-
-func GetNamespaceFromProviderID(providerID string) (string, error) {
-	if providerID == "" {
-		return "", errors.New("ProviderID is empty")
-	}
-
-	if !strings.HasPrefix(providerID, fmt.Sprintf("%s://", CloudProviderName)) {
-		return "", errors.New("ProviderID prefix is not valid")
-	}
-
-	providerIDParts := strings.Split(providerID, "/")
-	if len(providerIDParts) != 4 {
-		return "", errors.New("ProviderID is not valid")
-	}
-
-	return providerIDParts[2], nil
-}
